feat: gallery style banner
